Audio callback fixed.  Some clang compiler errors/warnings fixed. Some tsan warnings fixed.
Former-commit-id: ba7724d51febbef7deec781c989efc1ac5cdca5e [formerly 0349aeb4136b750290d9ca384e684235dcbbbcc1]
Former-commit-id: 76f98c79339969ebe77e90eceb1957173d9221bc

+++ mrViewer/src/core/aviImage.cpp
@@ -403,7 +403,7 @@ bool aviImage::test(const boost::uint8_t *data, unsigned len)
           memset( d+len, 0, AVPROBE_PADDING_SIZE );
           memcpy( d, data, len );
 
-          AVProbeData pd = { NULL, d, len, "video/MP2T" };
+          AVProbeData pd = { NULL, d, static_cast<int>(len), "video/MP2T" };
           AVInputFormat* ctx = av_probe_input_format(&pd, 1);
 
           delete [] d;
